+++ tests/uni.js
@@ -1,0 +1,5 @@
+"use strict";
+var greeting = "join room";
+var marker = "__tag";
+var o = {"plain": "v"};
+var tpl = `a${greeting}"b"`;